add 0792004 simple-io-2 experiment

+++ 0792004/simple-io-1/source/Exp_Homework.c
@@ -26,7 +26,7 @@
 void Exp_1_Homework_A(void)
 {
 	u16 sw;
-	u16 led = MIN_LED;
+	u8 led = MIN_LED;
 	u8 keyRight_pressed = FALSE;
 	u8 keyLeft_pressed = FALSE;
 
@@ -66,7 +66,7 @@ void Exp_1_Homework_A(void)
 void Exp_1_Homework_B(void)
 {
 	u16 sw;
-	u16 led = MIN_LED;
+	u8 led = MIN_LED;
 	u8 bar_sw = FALSE;	// TRUE = LED1 : FALSE = LED2
 	u8 keyR_pressed = FALSE;
 	u8 keyL_pressed = FALSE;
